feat(3293WH): GREEN — setup/upgrade self-verify via extracted health core
Extract checkHealth/reportHealthSummary into src/health.ts (network-free by
contract); wire both command tails to report + exit non-zero; parameterized
repairHint so post-upgrade failure never says 'run safeword upgrade'; demote
check to automatic-first in cli.mdx. REFACTOR folded in: firstFailureSection
+ selfVerify extraction to clear complexity caps.
